# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Integration with PHP Laravel",
        "description": "Integrate Swagger in Laravel application",
        "contact": {
            "email": "chandanmohanty11111@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/forgotpassword": {
            "post": {
                "summary": "forgotpassword",
                "description": "Send Mail to the respectice mail id for forget password link",
                "operationId": "186a73907a51e79b61ee87564453da94",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "we have mailed your password reset link to respective E-mail"
                    },
                    "404": {
                        "description": "we can not find a user with that email address"
                    }
                }
            }
        },
        "/api/auth/resetpassword": {
            "post": {
                "summary": "resetpassword",
                "description": "reset your password",
                "operationId": "b3e651e8d03d5635958aef7d32e69961",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "new_password",
                                    "confirm_password"
                                ],
                                "properties": {
                                    "new_password": {
                                        "type": "password"
                                    },
                                    "confirm_password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "we have mailed your password reset link to respective E-mail"
                    },
                    "404": {
                        "description": "we can not find a user with that email address"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Laravel bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}